+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 **Under development**
 
+- fix: secondary location model used same random seed in every parallel thread
+- feat: add a new method for attributing income to housholds using the bhepop2 package
 - fix: fixed special case in repairing ENTD for completely overlapping trips
 - feat: make it possible to disable the test run of MATSim before writing everything out
 - feat: check availability of open data sources for every PR